rag_api/requirements.txt
fastapi
uvicorn[standard]
opensearch-py
pydantic
python-dotenv
requests